pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "implicit-weight-field"
version = "0.1.0"
description = "Neural network compression using implicit weight field representations"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
    "tensorboard>=2.8.0",
    "transformers>=4.20.0",
    "tensorly>=0.7.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "hydra-core>=1.2.0",
    "wandb>=0.12.0",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]
experiments = [
    "torchattacks>=3.2.0",
    "foolbox>=3.3.0",
    "nvidia-ml-py>=11.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "models*", "compression*", "inference*", "evaluation*", "experiments*", "baselines*", "utils*", "visualization*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true